requirements.txt
# Core data handling
pandas>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.8.0

# Supabase integration
supabase>=2.0.0
python-dotenv>=1.0.0

# FastAPI for REST endpoints
fastapi>=0.104.0
uvicorn>=0.24.0

# HTTP requests
requests>=2.31.0

# Scheduling and background tasks
schedule>=1.2.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
